requirements.txt
aiogram~=3.13.1
pydantic~=2.9.2
pydantic_settings~=2.5.2
google-generativeai~=0.8.2
Pillow~=10.4.0